stream buckets content

+++ store/merkle_tree.go
@@ -66,7 +66,7 @@ func (content MerkleContent) CalculateHash() ([]byte, error) {
 type MerkleBucket struct {
 	contentList []MerkleContent
 	hash        []byte
-	bucketId    int
+	bucketId    int32
 }
 
 func (bucket MerkleBucket) CalculateHash() ([]byte, error) {
@@ -102,7 +102,7 @@ func PartitionMerkleTree(treeEpoch uint64, partitionId int) (*merkletree.MerkleT
 	bucketList := make([]MerkleBucket, partitionBuckets)
 
 	for i := range bucketList {
-		bucket := MerkleBucket{contentList: []MerkleContent{}, bucketId: i}
+		bucket := MerkleBucket{contentList: []MerkleContent{}, bucketId: int32(i)}
 		items := partition.Items(i, 0, int(treeEpoch))
 		for _, item := range items {
 			bucket.contentList = append(bucket.contentList, MerkleContent{key: item.Key, value: item.Value})
